List all Projects by owner. Dynamic project selection.

+++ contracts/ConstructionProject.sol
@@ -5,15 +5,17 @@ pragma solidity >=0.4.22 <0.7.0;
  * @dev Store & retreive value in a variable
  */
 contract ConstructionProject {
-
-    event WorkItemProgressChange(bytes32 indexed workItemId, WorkItemProgress workItemProgress);
+    event WorkItemProgressChange(
+        bytes32 indexed workItemId,
+        WorkItemProgress workItemProgress
+    );
 
     // Leistungsverzeichnis
     struct BOQ {
         bool exists;
         bytes32[] items;
     }
-    
+
     // Leistungsverzeichnis-Eintrag
     struct WorkItem {
         bytes32 id;
@@ -21,8 +23,8 @@ contract ConstructionProject {
         bytes32[] items;
         WorkItemProgress progress;
     }
-    
-    enum WorkItemProgress { INITIALIZED, STARTED, FINISHED, APPROVED }
+
+    enum WorkItemProgress {INITIALIZED, STARTED, FINISHED, APPROVED}
 
     address public generalContractor; // Rolle Generalunternehmer
     address public buildingContractor; // Rolle Bauherr
@@ -31,8 +33,7 @@ contract ConstructionProject {
     mapping(bytes32 => BOQ) BOQs; // Leistungsverzeichnisse mit Referenz auf Work Items
 
     mapping(bytes32 => WorkItem) workItems; // Leistungsverzeichnis-Einträge
-    mapping(bytes32 => bytes32) billingItems; // Work Items Referenz auf Abrechnungseinheiten  
-
+    mapping(bytes32 => bytes32) billingItems; // Work Items Referenz auf Abrechnungseinheiten
 
     modifier onlyGeneralContractor {
         require(msg.sender == generalContractor, 'No permission.');
@@ -53,17 +54,21 @@ contract ConstructionProject {
         subContractors[msg.sender] = true;
     }
 
-   /**
+    /**
      * @dev Adds a new Bill of Quantity
      * @param _BOQ The hash of the BOQ
      */
-    function addBOQ(bytes32 _BOQ) public onlyGeneralContractor returns (bool success) {
-        require(!BOQs[_BOQ].exists, "BOQ already specified.");
+    function addBOQ(bytes32 _BOQ)
+        public
+        onlyGeneralContractor
+        returns (bool success)
+    {
+        require(!BOQs[_BOQ].exists, 'BOQ already specified.');
         BOQs[_BOQ].items = new bytes32[](0);
         BOQs[_BOQ].exists = true;
         return true;
     }
-    
+
     /**
      * @dev Pushes a new billing item to the stack
      * @param _billingItem The hash of the billing item
@@ -78,7 +83,7 @@ contract ConstructionProject {
         // emit BillingItemCreated(billingUnitHash);
         return true;
     }
-    
+
     /**
      * @dev Pushes a new billing unit to the stack
      * @param _id The hash of the work item to be added
@@ -89,37 +94,40 @@ contract ConstructionProject {
         onlyGeneralContractor
         returns (bool success)
     {
-        require(workItems[_parent].id != 0x0 || BOQs[_parent].exists, "Parent is not specified.");
-        require(workItems[_id].id == 0x0, "Work Item is already specified."); // require(workItems[_id].progress < WorkItemProgress.INITIALIZED);
-        
+        require(
+            workItems[_parent].id != 0x0 || BOQs[_parent].exists,
+            'Parent is not specified.'
+        );
+        require(workItems[_id].id == 0x0, 'Work Item is already specified.'); // require(workItems[_id].progress < WorkItemProgress.INITIALIZED);
+
         workItems[_id] = WorkItem({
-            id: _id, 
-            parent: _parent, 
+            id: _id,
+            parent: _parent,
             items: new bytes32[](0),
             progress: WorkItemProgress.INITIALIZED
         });
-        if(BOQs[_parent].exists){
+        if (BOQs[_parent].exists) {
             BOQs[_parent].items.push(_id);
         } else {
             workItems[_parent].items.push(_id);
         }
         return true;
     }
-    
+
     /**
      * @dev Sets the progress of a work item and all of its subitems to a predefined state
      * @param _id The hash of the work item to be updated
      * @param _progress The state as uint to set the progress of the work item (0=INITIALIZED, 1=STARTED, 2=STOPPED, 3=FINISHED)
      */
-    function setWorkItemProgress(bytes32 _id, uint _progress)
+    function setWorkItemProgress(bytes32 _id, uint256 _progress)
         public
         returns (bool success)
     {
-        require(uint(workItems[_id].progress) < _progress);
+        require(uint256(workItems[_id].progress) < _progress);
         workItems[_id].progress = WorkItemProgress(_progress);
-        for (uint i=0; i<workItems[_id].items.length; i++) {
+        for (uint256 i = 0; i < workItems[_id].items.length; i++) {
             bytes32 itemId = workItems[_id].items[i];
-            if (uint(workItems[itemId].progress) < _progress) {
+            if (uint256(workItems[itemId].progress) < _progress) {
                 setWorkItemProgress(itemId, _progress);
             }
         }
